made NewItems.jsx dynamic
