+++ src/components/home/HotCollections.jsx
@@ -1,14 +1,15 @@
 import React, { useState, useEffect, useRef } from "react";
-import { Link } from "react-router-dom";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import axios from "axios";
+import HotCollectionsSkeleton from "./hot-collections/HotCollectionSkeleton";
+import HotCollection from "./hot-collections/HotCollection";
 
-const HotCollections = () => {
+const HotCollections = ({ sliderSettings }) => {
   const [hotCol, setHotCol] = useState([]);
   const [loading, setLoading] = useState(false);
-  const sliderRef = useRef(null);
+  
 
   async function getData() {
     const { data } = await axios.get(
@@ -23,108 +24,6 @@ const HotCollections = () => {
     getData();
   }, []);
 
-  useEffect(() => {
-  // Run after DOM paint
-  const resizeSlick = () => {
-    if (sliderRef.current && sliderRef.current.innerSlider) {
-      sliderRef.current.innerSlider.onWindowResized();
-    }
-  };
-
-  // First tick after mount
-  requestAnimationFrame(resizeSlick);
-
-  // Also after images load
-  window.addEventListener("load", resizeSlick);
-
-  return () => window.removeEventListener("load", resizeSlick);
-}, []);
-
-  function NextArrow({ onClick }) {
-    return (
-      <div
-        onClick={onClick}
-        className="custom-arrow next"
-        style={{
-          position: "absolute",
-          top: "50%",
-          right: "-25px",
-          transform: "translateY(-50%)",
-          zIndex: 2,
-          width: 40,
-          height: 40,
-          background: "#fff",
-          borderRadius: "50%",
-          boxShadow: "0 2px 6px rgba(0,0,0,0.2)",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          cursor: "pointer",
-        }}
-      >
-        <i className="fa fa-chevron-right" style={{ fontSize: 16 }} />
-      </div>
-    );
-  }
-
-  function PrevArrow({ onClick }) {
-    return (
-      <div
-        onClick={onClick}
-        className="custom-arrow prev"
-        style={{
-          position: "absolute",
-          top: "50%",
-          left: "-25px",
-          transform: "translateY(-50%)",
-          zIndex: 2,
-          width: 40,
-          height: 40,
-          background: "#fff",
-          borderRadius: "50%",
-          boxShadow: "0 2px 6px rgba(0,0,0,0.2)",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          cursor: "pointer",
-        }}
-      >
-        <i className="fa fa-chevron-left" style={{ fontSize: 16 }} />
-      </div>
-    );
-  }
-
-  const settings = {
-    dots: false,
-    arrows: true,
-    infinite: true,
-    speed: 500,
-    slidesToShow: 4,
-    slidesToScroll: 1,
-    nextArrow: <NextArrow />,
-    prevArrow: <PrevArrow />,
-    responsive: [
-      {
-        breakpoint: 1024,
-        settings: {
-          slidesToShow: 3,
-        },
-      },
-      {
-        breakpoint: 768,
-        settings: {
-          slidesToShow: 2,
-        },
-      },
-      {
-        breakpoint: 480,
-        settings: {
-          slidesToShow: 1,
-        },
-      },
-    ],
-  };
-
   return (
     <section id="section-collections" className="no-bottom">
       <div className="container">
@@ -135,78 +34,13 @@ const HotCollections = () => {
               <div className="small-border bg-color-2"></div>
             </div>
           </div>
-          <Slider ref={sliderRef} {...settings}>
+          <Slider {...sliderSettings}>
             {loading
               ? new Array(4).fill(0).map((_, index) => (
-                  <div key={index}>
-                    <div className="nft_coll">
-                      <div className="nft_wrap">
-                        <div
-                          className="skeleton-box"
-                          style={{
-                            width: "100%",
-                            height: "200px",
-                            borderRadius: "10px",
-                          }}
-                        />
-                      </div>
-                      <div className="nft_coll_pp">
-                        <div
-                          className="skeleton-box"
-                          style={{
-                            width: "50px",
-                            height: "50px",
-                            borderRadius: "50%",
-                          }}
-                        />
-                      </div>
-                      <div className="nft_coll_info">
-                        <div
-                          className="skeleton-box"
-                          style={{
-                            width: "70%",
-                            height: "16px",
-                            marginBottom: "8px",
-                          }}
-                        />
-                        <div
-                          className="skeleton-box"
-                          style={{ width: "40%", height: "14px" }}
-                        />
-                      </div>
-                    </div>
-                  </div>
+                  <HotCollectionsSkeleton key={index} />
                 ))
               : hotCol.map((collection) => (
-                  <div className="px-1" key={collection.id}>
-                    <div className="nft_coll">
-                      <div className="nft_wrap">
-                        <Link to={`/item-details/${collection.nftId}`}>
-                          <img
-                            src={collection.nftImage}
-                            className="lazy img-fluid"
-                            alt=""
-                          />
-                        </Link>
-                      </div>
-                      <div className="nft_coll_pp">
-                        <Link to={`/author/${collection.authorId}`}>
-                          <img
-                            className="lazy pp-coll"
-                            src={collection.authorImage}
-                            alt=""
-                          />
-                        </Link>
-                        <i className="fa fa-check"></i>
-                      </div>
-                      <div className="nft_coll_info">
-                        <Link to="/explore">
-                          <h4>{collection.title}</h4>
-                        </Link>
-                        <span>ERC-{collection.code}</span>
-                      </div>
-                    </div>
-                  </div>
+                  <HotCollection key={collection.id} collection={collection}/>
                 ))}
           </Slider>
         </div>
