+++ src/css/styles/style.css
@@ -16702,6 +16702,7 @@ img {
   animation: shimmer 2s infinite;
   content: "";
 }
+
 @-webkit-keyframes shimmer {
   100% {
     transform: translateX(100%);
