+++ src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useRef } from "react";
 import BrowseByCategory from "../components/home/BrowseByCategory";
 import HotCollections from "../components/home/HotCollections";
 import Landing from "../components/home/Landing";
@@ -7,6 +7,91 @@ import NewItems from "../components/home/NewItems";
 import TopSellers from "../components/home/TopSellers";
 
 const Home = () => {
+  function NextArrow({ onClick }) {
+    return (
+      <div
+        onClick={onClick}
+        className="custom-arrow next"
+        style={{
+          position: "absolute",
+          top: "50%",
+          right: "0",
+          transform: "translateY(-50%)",
+          zIndex: 2,
+          width: 40,
+          height: 40,
+          background: "#fff",
+          borderRadius: "50%",
+          boxShadow: "0 2px 6px rgba(0,0,0,0.2)",
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          cursor: "pointer",
+        }}
+      >
+        <i className="fa fa-chevron-right" style={{ fontSize: 16 }} />
+      </div>
+    );
+  }
+
+  function PrevArrow({ onClick }) {
+    return (
+      <div
+        onClick={onClick}
+        className="custom-arrow prev"
+        style={{
+          position: "absolute",
+          top: "50%",
+          left: "0",
+          transform: "translateY(-50%)",
+          zIndex: 2,
+          width: 40,
+          height: 40,
+          background: "#fff",
+          borderRadius: "50%",
+          boxShadow: "0 2px 6px rgba(0,0,0,0.2)",
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          cursor: "pointer",
+        }}
+      >
+        <i className="fa fa-chevron-left" style={{ fontSize: 16 }} />
+      </div>
+    );
+  }
+
+  const sliderSettings = {
+    dots: false,
+    arrows: true,
+    infinite: true,
+    speed: 500,
+    slidesToShow: 4,
+    slidesToScroll: 1,
+    nextArrow: <NextArrow />,
+    prevArrow: <PrevArrow />,
+    responsive: [
+      {
+        breakpoint: 1024,
+        settings: {
+          slidesToShow: 3,
+        },
+      },
+      {
+        breakpoint: 768,
+        settings: {
+          slidesToShow: 2,
+        },
+      },
+      {
+        breakpoint: 480,
+        settings: {
+          slidesToShow: 1,
+        },
+      },
+    ],
+  };
+
   useEffect(() => {
     window.scrollTo(0, 0);
   }, []);
@@ -17,8 +102,8 @@ const Home = () => {
         <div id="top"></div>
         <Landing />
         <LandingIntro />
-        <HotCollections />
-        <NewItems />
+        <HotCollections sliderSettings={sliderSettings} />
+        <NewItems sliderSettings={sliderSettings} />
         <TopSellers />
         <BrowseByCategory />
       </div>
